backend/pyproject.toml
[project]
name = "point-source"
version = "0.1.0"
description = "Pydantic AI Production Ready Template"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pydantic-ai>=1.13.0",
    "fastapi>=0.121.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "SQLAlchemy>=2.0.44",
    "alembic>=1.17.1",
    "asyncpg>=0.31.0",
    "greenlet>=3.2.4",
    "redis>=7.1.0",
    "fastapi-guard>=4.2.1",
    "python-jose[cryptography]>=3.3.0",
    "starlette-admin>=0.15.1",
    "bcrypt>=5.0.0",
    "itsdangerous>=2.2.0",
    "typer>=0.20.0",
    "pgvector>=0.4.2",
    "datasets>=4.5.0",
    "logfire[fastapi,litellm,redis,sqlalchemy]>=4.19.0",
    "aioitertools>=0.13.0",
    "sentence-transformers>=5.2.2",
    "mlx>=0.30.6",
    "mlx-lm>=0.29.1",
    "mlx-metal>=0.30.6",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "cookiecutter>=2.6.0",
    "cookiecutter-data-science>=2.3.0",
    "httpx>=0.27.2",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "nox>=2025.5.1",
    "pre-commit>=4.3.0",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "pydantic-evals[logfire]>=1.16.0",
    "transformers>=4.57.6",
    "numpy>=2.4.1",
    "litellm>=1.81.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "evaluation"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
exclude = [
    "alembic",
    ".venv",
    "venv",
    "__pycache__",
]
